+++ accounts/views.py
@@ -1,0 +1,57 @@
+from django.shortcuts import render
+from django.views.generic import View
+from .forms import UserForm, SignUpForm
+from django.contrib.auth import authenticate, login
+
+
+class UserFormView(View):
+    form_class = UserForm
+    template_name = 'accounts/login_form.html'
+
+    # display blank form
+    def get(self, request):
+        form = self.form_class(None)
+        return render(request, self.template_name, {'form': form})
+
+    # process form data
+    def post(self, request):
+        form = self.form_class(request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)
+
+            # clean (normalized) data
+            # username = form.cleaned_data['username']
+            email = form.cleaned_data['email']
+            password = form.cleaned_data['password']
+            user.set_password(password)
+            user.save()
+
+            # returns User objects if credentials are correct
+            user = authenticate(email, password)
+
+            if user is not None:
+                if user.is_active:
+                    login(request, user)
+                    return redirect('music:index')
+        return render(request, self.template_name, {'form': form})
+
+
+class UserRegView(View):
+    form_class = SignUpForm
+    template_name = 'accounts/registration_form.html'
+
+    # display blank form
+    def get(self, request):
+        form = self.form_class(None)
+        return render(request, self.template_name, {'form': form})
+
+    def signup(self, request):
+        form = SignUpForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.is_active = False
+            user.save()
+            return redirect('music:index')
+
+        return render(request, self.template_name, {'form': form})
