+++ music/models.py
@@ -1,14 +1,5 @@
-from __future__ import unicode_literals
-
 from django.db import models
 from django.core.urlresolvers import reverse
-
-from django.contrib.auth.models import User
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-
-
-# Create your models here.
 
 
 class Album(models.Model):
@@ -28,21 +19,8 @@ class Song(models.Model):
     album = models.ForeignKey(Album, on_delete=models.CASCADE)
     file_type = models.CharField(max_length=10)
     song_title = models.CharField(max_length=250)
+    song_file = models.FileField()
     is_favorite = models.BooleanField(default=False)
 
     def __str__(self):
         return self.song_title
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    location = models.CharField(max_length=30, blank=True)
-    birth_date = models.DateField(null=True, blank=True)
-    email_confirmed = models.BooleanField(default=False)
-
-
-@receiver(post_save, sender=User)
-def update_user_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
-    instance.profile.save()
