--- music/static/music/style.css
@@ -1,11 +1,0 @@
-body{
-    background: #fff url('images/background.png');
-}
-
-.navbar{
-    border-radius:0;
-}
-
-.navbar-brand{
-    font-family:'Satisfy','cursive';
-}
