--- music/templates/music/account_activation_email.html
@@ -1,7 +1,0 @@
-{% autoescape off %}
-Hi {{ user.username }},
-
-Please click on the link below to confirm your registration:
-
-http://{{ domain }}{% url 'activate' uidb64=uid token=token %}
-{% endautoescape %}
