--- music/templates/music/account_activation_invalid.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-  <p>The confirmation link was invalid, possibly because it has already been used.</p>
-{% endblock %}
