--- music/templates/music/account_activation_sent.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-  <p>Please confirm your email address to complete the registration. </p>
-{% endblock %}
