--- music/templates/music/index.html
@@ -1,119 +1,0 @@
-{% extends 'music/base.html' %}
-
-{% block body %}
-    <!-- {% if all_albums %}
-        <h3>Here are all my Albums:</h3>
-        <ul>
-            {% for album in all_albums %}
-            <li><a href="{% url 'music:detail' album.id %}">{{ album.album_title }}</a></li>
-            {% endfor %}
-        </ul>
-    {% else %}
-        <h3>You don't have any albums</h3>
-    {% endif %} -->
-    <div class="albums-container container-fluid">
-
-            <!-- Albums -->
-            <div class="row">
-                <div class="col-sm-12">
-                    <h3>My Albums</h3>
-                </div>
-                {% if all_albums %}
-                    {% for album in all_albums %}
-                        <div class="col-sm-4 col-lg-2">
-                            <div class="thumbnail">
-                                <a href="{% url 'music:detail' album.id %}">
-                                    <img src="{{ album.album_logo.url }}" class="img-responsive">
-                                </a>
-                                <div class="caption">
-                                    <h2>{{ album.album_title }}</h2>
-                                    <h4>{{ album.artist }}</h4>
-        
-                                    <!-- View Details -->
-                                    <a href="{% url 'music:detail' album.id %}" class="btn btn-primary btn-sm" role="button">View Details</a>
-        
-                                    <!-- Delete Album -->
-                                    <form action="{% url 'music:album-delete' album.id %}" method="post" style="display: inline;">
-                                        {% csrf_token %}
-                                        <input type="hidden" name="album_id" value="{{ album.id }}" />
-                                        <button type="submit" class="btn btn-default btn-sm">
-                                            <span class="glyphicon glyphicon-trash"></span>
-                                        </button>
-                                    </form>
-        
-                                    <!-- Favorite Album -->
-                                    <a href="#" class="btn btn-default btn-sm btn-favorite" role="button">
-                                        <span class="glyphicon glyphicon-star"></span>
-                                    </a>
-        
-                                </div>
-                            </div>
-                        </div>
-                        {% cycle '' '' '' '' '' '<div class="clearfix visible-lg"></div>' %}
-                    {% endfor %}
-                {% else %}
-                    <div class="col-sm-12">
-                        <br>
-                        <a href="{% url 'music:album-add' %}">
-                            <button type="button" class="btn btn-success">
-                                <span class="glyphicon glyphicon-plus  {% if album.is_favorite %}active{% endif %}"></span>&nbsp; Add an Album
-                            </button>
-                        </a>
-                    </div>
-                {% endif %}
-            </div>
-        
-            <!-- If user searches and there are songs -->
-            {% if songs %}
-                <div class="row">
-                    <div class="col-sm-12">
-                        <h3>Songs</h3>
-                    </div>
-                    <div class="col-sm-12">
-                        <div class="panel panel-default">
-                            <div class="panel-body">
-                                <table class="table">
-                                    <thead>
-                                        <tr>
-                                            <th>Song Title</th>
-                                            <th>Artist</th>
-                                            <th>Audio File</th>
-                                            <th>Album</th>
-                                            <th>Favorite</th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        {% for song in songs %}
-                                            <tr>
-                                                <td>{{ song.song_title }}</td>
-                                                <td>{{ song.album.artist }}</td>
-                                                <td>
-                                                    <a target="_blank" href="#">
-                                                        <button type="button" class="btn btn-success btn-xs">
-                                                            <span class="glyphicon glyphicon-play"></span>&nbsp; Play
-                                                        </button>
-                                                    </a>
-                                                </td>
-                                                <td>
-                                                    <a href="{% url 'music:detail' song.album.id %}">
-                                                        <img src="{{ song.album.album_logo.url }}" class="img-responsive" style="width: 20px; float: left; margin-right: 10px;" />
-                                                    </a>
-                                                    <a href="{% url 'music:detail' song.album.id %}">{{ song.album.album_title }}</a>
-                                                </td>
-                                                <td>
-                                                    <a href="#" class="btn-favorite"><span class="glyphicon glyphicon-star {% if song.is_favorite %}active{% endif %}"></span></a>
-                                                </td>
-                                            </tr>
-                                        {% endfor %}
-                                    </tbody>
-                                </table>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            {% endif %}
-        
-        </div>
-
-
-{% endblock %}
