+++ music/urls.py
@@ -5,14 +5,10 @@ app_name = 'music'
 
 urlpatterns = [
     url(r'^$', views.IndexView.as_view(), name='index'),
-    url(r'^register/$', views.UserRegView.as_view(), name='register'),
-    url(r'^login/$', views.UserFormView.as_view(), name='login'),
     url(r'^(?P<pk>[0-9]+)/$', views.DetailView.as_view(), name='detail'),
-    url(r'^account_activation_sent/$', views.account_activation_sent, name='account_activation_sent'),
-    url(r'^activate/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
-        views.activate, name='activate'),
     #/music/album/add/
     url(r'^album/add/$', views.AlbumCreate.as_view(), name='album-add'),
+    url(r'^album/song/add/$', views.SongCreate.as_view(), name='song-add'),
     #/music/album/2/
     url(r'^album/(?P<pk>[0-9]+)/$', views.AlbumUpdate.as_view(), name='album-update'),
     #/music/album/2/delete/
