+++ music/views.py
@@ -1,19 +1,9 @@
 from django.views import generic
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.shortcuts import render, redirect
-from django.contrib.auth import authenticate, login
 from django.views.generic import View
 from django.core.urlresolvers import reverse_lazy
-from .models import Album
-from .forms import UserForm, SignUpForm
-
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode
-from django.template.loader import render_to_string
-from .tokens import account_activation_token, account_activation_token
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode
-from django.template.loader import render_to_string
+from .models import Album, Song
 
 
 class IndexView(generic.ListView):
@@ -44,63 +34,16 @@ class AlbumDelete(DeleteView):
     success_url = reverse_lazy('music:index')
 
 
-class UserFormView(View):
-    form_class = UserForm
-    template_name = 'music/login_form.html'
-
-    # display blank form
-    def get(self, request):
-        form = self.form_class(None)
-        return render(request, self.template_name, {'form': form})
-
-    # process form data
-    def post(self, request):
-        form = self.form_class(request.POST)
-
-        if form.is_valid():
-            user = form.save(commit=False)
-
-            # clean (normalized) data
-            # username = form.cleaned_data['username']
-            email = form.cleaned_data['email']
-            password = form.cleaned_data['password']
-            user.set_password(password)
-            user.save()
-
-            # returns User objects if credentials are correct
-            user = authenticate(email, password)
-
-            if user is not None:
-                if user.is_active:
-                    login(request, user)
-                    return redirect('music:index')
-        return render(request, self.template_name, {'form': form})
+class SongCreate(CreateView):
+    model = Song
+    fields = ['song_title', 'song_file', 'file_type', 'is_favorite']
 
 
-class UserRegView(View):
-    form_class = SignUpForm
-    template_name = 'music/registration_form.html'
+class SongUpdate(UpdateView):
+    model = Song
+    fields = ['song_title', 'song_file', 'file_type', 'is_favorite']
 
-    # display blank form
-    def get(self, request):
-        form = self.form_class(None)
-        return render(request, self.template_name, {'form': form})
 
-    def signup(self, request):
-        form = SignUpForm(request.POST)
-        if form.is_valid():
-            user = form.save(commit=False)
-            user.is_active = False
-            user.save()
-            current_site = get_current_site(request)
-            subject = 'Activate Your MySite Account'
-            message = render_to_string('account_activation_email.html', {
-                'user': user,
-                'domain': current_site.domain,
-                'uid': urlsafe_base64_encode(force_bytes(user.pk)),
-                'token': account_activation_token.make_token(user),
-            })
-            user.email_user(subject, message)
-            return redirect('music:account_activation_sent')
-
-        return render(request, self.template_name, {'form': form})
+class SongDelete(DeleteView):
+    model = Song
+    success_url = reverse_lazy('music:index')
