+++ templates/accounts/login_form.html
@@ -1,4 +1,4 @@
-{% extends 'music/base.html' %}
+{% extends 'base.html' %}
 {% block title %}Login{% endblock %}
 
 {% block body %}
@@ -15,12 +15,12 @@
                     {% endif %}
                     <form class="form-horizontal" role="form" action="" method="post" enctype="multipart/form-data">
                         {% csrf_token %}
-                        {% include 'music/form_template.html' %}
+                        {% include 'form_template.html' %}
                         <div class="form-group">
                             <div class="col-sm-offset-2 col-sm-10">
                                 <button type="submit" class="btn btn-success">Submit</button>
                             </div>
-                        </div> 
+                        </div>
                     </form>
                 </div>
             </div>
