+++ templates/music/album_form.html
@@ -1,5 +1,6 @@
-{% extends 'music/base.html' %}
-{% block title %}Register{% endblock %}
+{% extends 'base.html' %}
+{% block title %}Add a New Album{% endblock %}
+{% block albums_active %}active{% endblock %}
 
 {% block body %}
 <div class="container-fluid">
@@ -9,18 +10,18 @@
         <div class="col-sm-12 col-md-7">
             <div class="panel panel-default">
                 <div class="panel-body">
-                    <h3>Create New Account</h3>
+                    <h3>Add a New Album</h3>
                     {% if error_message %}
                         <p><strong>{{ error_message }}</strong></p>
                     {% endif %}
                     <form class="form-horizontal" role="form" action="" method="post" enctype="multipart/form-data">
                         {% csrf_token %}
-                        {% include 'music/form_template.html' %}
+                        {% include 'form_template.html' %}
                         <div class="form-group">
                             <div class="col-sm-offset-2 col-sm-10">
                                 <button type="submit" class="btn btn-success">Submit</button>
                             </div>
-                        </div> 
+                        </div>
                     </form>
                 </div>
             </div>
