+++ templates/music/detail.html
@@ -1,4 +1,4 @@
-{% extends 'music/base.html' %}
+{% extends 'base.html' %}
 {% block title %}Album Details{% endblock %}
 {% block body %}
     <div class="container-fluid songs-container">
@@ -19,24 +19,24 @@
                     </div>
                 </div>
             </div>
-            
+
             <!-- Right Song Info -->
             <div class="col-sm-8 col-md-9">
-    
+
                 <ul class="nav nav-pills" style="margin-bottom: 10px;">
                     <li role="presentation" class="active"><a href="{% url 'music:detail' album.id %}">View All</a></li>
                     <li role="presentation"><a href="#">Add New Song</a></li>
                 </ul>
-    
+
                 <div class="panel panel-default">
                     <div class="panel-body">
-    
+
                         <h3>All Songs</h3>
-    
+
                         {% if error_message %}
                             <p><strong>{{ error_message }}</strong></p>
                         {% endif %}
-    
+
                         <table class="table">
                             <thead>
                             <tr>
@@ -73,14 +73,14 @@
                             {% endfor %}
                             </tbody>
                         </table>
-    
+
                     </div>
                 </div>
-            
+
             </div>
-            
+
         </div>
-            
+
      </div>
-    
+
 {% endblock %}
